+++ src/Transports/BootsEverywhereWalkers.java
@@ -13,10 +13,10 @@ public class BootsEverywhereWalkers extends GroundTransport {
     }
 
     @Override
-    public int GetRacingTime(Integer distance) {
+    public double GetRacingTime(double distance) {
 
-        int restCount = distance / speed;
-        int totalTime = 0;
+        int restCount = (int) distance / speed;
+        double totalTime = 0;
 
         for (int i = 0; i < restCount; i++) {
