+++ src/Transports/Transport.java
@@ -2,13 +2,10 @@ package Transports;
 
 public abstract class Transport {
 
-
     public String name() { return "No name"; }
 
     public abstract boolean CanTakePart(RaceType type);
 
-    ////////////////////////////
-    public int GetRacingTime(Integer distance) { return 0; }
-
+    public double GetRacingTime(double distance) { return 0; }
 
 }
